+++ examples/java-server/src/test/java/com/twitter/hello/server/HelloWorldServerStartupTest.java
@@ -1,5 +1,8 @@
 package com.twitter.hello.server;
 
+import com.google.common.collect.ImmutableMap;
+import com.google.inject.Stage;
+
 import org.junit.Test;
 
 import com.twitter.inject.server.EmbeddedTwitterServer;
@@ -7,7 +10,10 @@ import com.twitter.inject.server.EmbeddedTwitterServer;
 public class HelloWorldServerStartupTest {
 
     private EmbeddedTwitterServer server =
-            new EmbeddedTwitterServer(new HelloWorldServer());
+        new EmbeddedTwitterServer(
+            new HelloWorldServer(),
+            ImmutableMap.of(),
+            Stage.PRODUCTION);
 
     @Test
     public void testServerStartup() {