+++ http/src/main/scala/com/twitter/finatra/http/JavaHttpServer.scala
@@ -1,3 +1,0 @@
-package com.twitter.finatra.http
-
-abstract class JavaHttpServer extends HttpServer