+++ jackson/src/main/scala/com/twitter/finatra/json/FinatraObjectMapper.scala
@@ -166,7 +166,7 @@ case class FinatraObjectMapper(
     }
   }
 
-  def registerModule(module: Module) = {
+  def registerModule(module: Module): ObjectMapper = {
     objectMapper.registerModule(module)
   }
 }