+++ thrift/src/main/scala/com/twitter/finatra/thrift/internal/routing/NullThriftService.scala
@@ -1,0 +1,5 @@
+package com.twitter.finatra.thrift.internal.routing
+
+import com.twitter.scrooge.ThriftService
+
+object NullThriftService extends ThriftService